removed new_db_manager and switched to db_manager
