+++ agenta-backend/agenta_backend/routers/evaluation_router.py
@@ -44,7 +44,7 @@ from agenta_backend.utils.common import engine, check_access_to_app
 from agenta_backend.services.db_manager import query, get_user_object
 from agenta_backend.models.db_models import EvaluationDB, EvaluationScenarioDB
 from agenta_backend.config import settings
-from agenta_backend.services import new_db_manager
+from agenta_backend.services import db_manager
 from agenta_backend.models import converters
 from agenta_backend.services import results_service
 
@@ -97,7 +97,7 @@ async def create_evaluation(
                 {"detail": error_msg},
                 status_code=400,
             )
-        app_ref = await new_db_manager.fetch_app_by_id(app_id=payload.app_id)
+        app_ref = await db_manager.fetch_app_by_id(app_id=payload.app_id)
 
         if app_ref is None:
             raise HTTPException(status_code=404, detail="App not found")
