+++ agenta-backend/agenta_backend/routers/user_profile.py
@@ -1,6 +1,6 @@
 import os
 from agenta_backend.utils.common import engine
-from agenta_backend.services.db_manager import UserDB
+from agenta_backend.models.db_models import UserDB
 from fastapi import APIRouter, HTTPException, Depends
 from agenta_backend.models.api.user_models import User
